py3/requirements.txt
backports-abc
beautifulsoup4
bs4
future
futures
markdown
markdown2
pyparsing
python-dateutil
PyYAML
singledispatch
six
tornado
gitpython
requests
prettierfier
jsonpickle
#weasyprint
https://github.com/Kozea/WeasyPrint/archive/master.zip
